chore: refactoring the refactor
