+++ scripts/scripts.js
@@ -4,7 +4,7 @@ import {
   loadHeader,
   loadFooter,
   decorateButtons,
-  decorateIcons,
+  decorateIcon,
   decorateSections,
   decorateBlocks,
   decorateTemplateAndTheme,
@@ -53,6 +53,92 @@ function addQuickNav() {
   }
 }
 
+const ICONS_CACHE = {};
+
+/**
+ * Attempt to replace <img> with <svg><use> to allow styling based on use of current color
+ * @param {icon} icon <img> element
+ */
+async function spriteIcon(icon) {
+  // Prepare the inline sprite
+  let svgSprite = document.getElementById('franklin-svg-sprite');
+  if (!svgSprite) {
+    const div = document.createElement('div');
+    div.innerHTML = '<svg xmlns="http://www.w3.org/2000/svg" id="franklin-svg-sprite" style="display: none"></svg>';
+    svgSprite = div.firstElementChild;
+    document.body.append(div.firstElementChild);
+  }
+
+  const { iconName } = icon.dataset;
+  if (!ICONS_CACHE[iconName]) {
+    try {
+      const response = await fetch(icon.src);
+      // cowardly refusing to load large icons
+      if (response.contentLength > 10240) {
+        ICONS_CACHE[iconName] = { };
+        return;
+      }
+      if (!response.ok) {
+        return;
+      }
+
+      const svg = await response.text();
+      const parser = new DOMParser();
+      const svgDOM = parser.parseFromString(svg, 'image/svg+xml');
+      const svgElem = svgDOM.querySelector('svg');
+
+      // only sprite icons that use currentColor
+      if (svg.toLowerCase().includes('currentcolor')) {
+        const symbol = document.createElementNS('http://www.w3.org/2000/svg', 'symbol');
+        symbol.id = `icons-sprite-${iconName}`;
+        symbol.setAttribute('viewBox', svgElem.getAttribute('viewBox'));
+        while (svgElem.firstChild) symbol.append(svgElem.firstChild);
+        ICONS_CACHE[iconName] = {
+          symbol,
+        };
+        svgSprite.append(symbol);
+      } else {
+        icon.alt = svgElem.querySelector('title') ? svgElem.querySelector('title').textContent : iconName;
+        ICONS_CACHE[iconName] = { };
+      }
+    } catch (error) {
+      // eslint-disable-next-line no-console
+      console.error(error);
+    }
+  }
+
+  if (document.getElementById(`icons-sprite-${iconName}`)) {
+    const span = icon.closest('span.icon');
+    if (span) span.innerHTML = `<svg xmlns="http://www.w3.org/2000/svg"><use href="#icons-sprite-${iconName}"/></svg>`;
+  }
+}
+
+/**
+ * Add intersection observer to all icons in an element, to sprite them if possible
+ * @param {Element} element element that contains icons
+ * @param {string} prefix prefix for icon names
+ */
+
+// eslint-disable-next-line import/prefer-default-export
+export function spriteIcons(element, prefix = '') {
+  const icons = [...element.querySelectorAll('span.icon')];
+  icons.forEach((span) => {
+    if (!span.firstElementChild) decorateIcon(span, prefix);
+    const img = span.querySelector('img');
+    if (img) {
+      const io = new IntersectionObserver((entries) => {
+        entries.forEach((entry) => {
+          if (entry.isIntersecting) {
+            spriteIcon(img);
+            io.disconnect();
+          }
+        });
+      });
+      io.observe(img);
+    }
+  });
+}
+
 function buildHeroBlock(main) {
   const h1 = main.querySelector('h1');
   const picture = main.querySelector('picture');
@@ -98,7 +184,7 @@ function buildAutoBlocks(main) {
 function decorateMain(main) {
   // hopefully forward compatible button decoration
   decorateButtons(main);
-  decorateIcons(main);
+  spriteIcons(main);
   buildAutoBlocks(main);
   decorateSections(main);
   decorateBlocks(main);
